backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
groq
pydantic
pydantic-settings
python-dotenv
python-multipart
pdfplumber
python-pptx
